build: update mocha for retries and audit fix

+++ spec/api-app-spec.ts
@@ -1516,6 +1516,7 @@ describe('app module', () => {
         // to the Docker invocation allows the syscalls that Chrome needs, but
         // are probably more permissive than we'd like.
         this.skip();
+        return;
       }
       fs.unlink(socketPath, () => {
         server = net.createServer();
@@ -1527,13 +1528,17 @@ describe('app module', () => {
     afterEach(done => {
       if (appProcess != null) appProcess.kill();
 
-      server.close(() => {
-        if (process.platform === 'win32') {
-          done();
-        } else {
-          fs.unlink(socketPath, () => done());
-        }
-      });
+      if (server) {
+        server.close(() => {
+          if (process.platform === 'win32') {
+            done();
+          } else {
+            fs.unlink(socketPath, () => done());
+          }
+        });
+      } else {
+        done();
+      }
     });
 
     describe('when app.enableSandbox() is called', () => {
